requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
